mysql/check_records/python/requirements.txt
python-dotenv
mysql.connector
